Fixed the DamageFriendly behavior
It was, due to a coding flaw, marking everything with DamageFriendly = true.    It's much better if specific types of weapons have that value manually set.  Otherwise default should be false.

GetDamageTable function entirely replaced with data stored on the weapon now - no more repeated function call every time a weapon fires.   Will keep the GetDamageTable function for now just in case something pops up - but otherwise it's deprecated.

INTELCHECKS - add reference to all the intel checks that can be made to the engine so that it's handy.

Resized UEL0108 - the UEF Crusher T1 tank from 4DC.

OPTIMIZATION - created priority functions that will shut off the lower end aircraft builders and thus removed the need for a builder condition that would have been checked for the rest of the game.

GETLISTOFUNITS - revised all  uses of it to insure proper use of the IsIdle and IsBuilt parameters.   IsIdle is a bit of a misnomer since it's possible to be Idle but actually be in a platoon.  This should generally always be FALSE - the other parameter just insures that it doesn't include units which are being built.