Engineer Callbacks
These fixes insure that whenever an engineer is transferred, it's callbacks are rebuilt 

Data Reduction in Shields - no longer carry the Brain value around as part of the entity.

Refined the PickEnemy logic so that we can identify the location of an opponent right from the start of the game.    This allows the immediate creation of an attack plan - and allows us to determine if a base position can reach the Attack Plan goal by LAND or AMPHIBIOUS.  To be used by the logic which picks a base layout for a new base position.   This in turn will allow LOUD to use a land based production profile on water maps when he can reach his enemy via land.