SIZESPHERE - A complete review on how SizeSphere works in an airunit blueprint
All units have had SizeSphere reset to a value that is related to airspeed.   This insures hitting air units in a reliable and predictable fashion.    Bad news for the air units in general since they were enjoying a high degree of safety from non-tracking weapons (SAMS).    In general, the SizeSphere should be set to a value of .1 of MaxAirspeed.   This insures that the unit is potentially hitable at that speed.   If the value is lower, the chance of hitting drops at an astounding rate.