LUA - CDR not upgrading in vanilla
Found a bug where the ACU would not enhance - this was realistically a problem in the way the conditions monitor treated conditions with only one parameter - it would try to test them as if they were location based conditions - which of course they would always fail.   Left some code tags in there for now, and added a LOG entry that fires when a CDR does an enhancement.