LUA - a very rare crash bug - fixed
In the rare event that an engineer would be stuck in WatchForNotBuilding, and then would die during the processing of that event - AND - the garbage collection cycle was collecting the engineer BUT not the now dead platoon - you  would crash.    The real culprit was really that the PBC function was being forked off the platoon - and not the engineer - which would allow PBC to execute even if the engineer was actually dead - and hence, our crash.   Refactored all the calls to PBC to insure that they all fork off the engineer rather than the platoon.