LUA - BUFFS issues
This is a long standing issue that I think is finally solved.   Some BUFFS launch threads - either because they have durations - or they have visual effects.   Either way - if the buff was reapplied, these threads stacked up in the trashbag - never to be relieved - eventually causing a stack overflow or massive slowdowns.