LUA - Adjacency Buffs reduced for T3 Power and Mass Fabs on very large structures
This is intended to address the magnitude of the bonuses that can be gotten on things like the Experimental Gantries of BrewLAN, where sometimes the discounts can be more than the economic unit actually produces.
Still need to find a way to cancel adjacency bonuses when a mass fab is turned off.