AI - Unit restrictions, Spawnwave
The restricted units tables were being rebuilt each time a unit was about to be constructed.  Now it's controlled by a global flag that is set once that has been done.   In order to get the restrictions to be rebuilt again - the global flag will have to be reset when any changes to unit restrictions are implemented during a game.
The AIx Spawnwave was not functioning properly when unit restrictions were enabled - this is now fixed and inline with the restrictions work above.   The spawnwave is now also making use of the AI's NeedTransports flag, which will control if any transports are included as part of the wave.