AI - Intel Changes
Lots of work here - primarily this - As an intel unit is moving to it's goal, check if passing by other intel points, and update them as we do - installed for Air, Land and Naval scouts.
PARSEINTEL - revised the collection of AIR and NAVAL intel to include all such units - not just the MOBILE ones.   This gives the AI a better indication of activity on these layers.
Threat decay rate has been reduced so that when an intel point is reported, but no units are seen, the intel value will only be reduced by 10% each time, rather than 20%.
Also added a check for intel 'aging' - so that when DisplayIntelPoints is turned on, it will also report the average age of both Hi and Low priority tables.
RATIOS - the maximum reported AIR. LAND and NAVAL ratios have been increased to 100 - this helps me in seeing the ratio movement from the AI perspective in greater detail.