AI - Distress Monitor for Bases
A tune up of Distress and Alert monitoring for bases
Bases will only send a LAND response if they can meet 60% of the threat to the base.
When responses are sent out, only an equal amount of threat will be sent out in any one pass - leaving some reserve for any additional responses.
As units are sent out - the units furthest away will be sent first - with slight delays for those units which are closer - this 'tightens' the response - and spreads the processing out over several ticks.
Rather than send all the units to the identical point of the threat - they're sent to random positions near the goal to minimize traffic issues
The Base Monitor will recognize the end of a threat better - and return the units to the base more quickly.
Units will now position themselves at the rally points closest to the threat when they cannot respond normally - and returning units will likewise return to the rally points closest to the threat
The delay between alert responses will grow more slowly and cap out much sooner (maximum delay between responses is now much lower).