AI - TRANSPORTS - Fixed some long-standing management issues
There were several niggling issues with transports.  Created a full logging suite with a switch in AIBrain to turn it on and off.   Fixed the Assignment/InUse flags to work properly when transports were RTB during assignment.
This was initially made clear when we saw an engineer start a base from kilometres away, after his transport decided to abort on him - which in turn had been caused because the transport got moved into the Transport Pool before properly exiting from the Refuel Pool.
Found some sequencing optimizations along the way.