LUA - Some revisions to AI and applications of cheat
With regards to recent changeover to a single AI multiplier, I took some time to review how the multiplier is applied - so that the multiplier, while it still affects mass/energy output when it is set to less than 1 - the other cheats (HP, Intel, etc) are NOT affected when it set below 1.
Also - removed the Non-Cheat buildergroup in ALL the base templates.