LUA - AttackForceAI_Bomber
Introduced a custom behavior applied to bombers.  This allows some experimentation with 'secondary' target bombing.   This is, in effect, similar to the manual split-attack that humans can use - splitting the bombers up, and applying a certain number of the bombers first to Shields and AA - before the primary target.
Also - repaired an issue with platoon merging