LUA - MovePlatoon function
Decreased the cycling period from 15 to 10 ticks as it was responsible for a meshing issue in the RTB function - causing some platoons to think they were idle because the MovePlatoon thread had not yet issued a new movement order - having completed the most recent stage.
In review - I was able to eliminate the use of the nocmdactive protocol, as in the places where it was used, the 'stuck' count would do the same job.