UNITS - Cybran T1 Gunship
The Cybran T1 gunship is an outlier amongst gunships and T1 air units overall.   To bring it back into the same realm as other gunships, it airspeed was dropped to 9, health was lowered to 345 and the costs were slightly nudged up.  Fuel time was lowered slightly.
As for the weapons, some randomness was added and turret speed was normalized.