LUA - Base Mode - Land/Amphibious
One of the problems with water maps is getting the AI to understand that he doesn't always have to produce Amphibious units - in fact - often maps have a LAND solution to the current target goal.   These changes allow him to see that - and produce units - and build platoons - accordingly - and on a per base level.  The one clawback to this process - we now have to load both land and amphibious builders and use them interactively - a small performance hit.