LUA - Pathfinding work
In order to expand upon the support for Land/Amphibious modes at bases - we needed to expand some of the data returned by pathfinding operations - so all pathfinding operations will now return a cost in addition to the path.
Additional changes have also been made to pathfinding in general, to incorporate changes so that amphibious pathfinding can tell is a node is in water - or not - and have an altered cost - and use a secondary threat type (ie - using sub threat versus surface when underwater).
A whole bunch of whitespace removal.