LUA - Fix for Factory being added to wrong base
Sometimes, an engineer in RTB will stop to assist another engineer - especially if he happens to be passing thru a base.   In this bug - the naval base engineer was building a new factory - BUT - when the build was complete - it was assigned to the base that the assisting engineer came from - WRONG - the base is now assigned when the build starts - and comes from the engineer that starts it - not the one that finishes.