LUA - INTEL study
In an effort to assess the IMAP Intel data - we discover that ARTILLERY just plain does not register.   And we discover that amphibious units go from being land threat when on land, to naval threat when on water.
This necessitates some visual changes to the intel logging routines.
Some data reduction is ParseIntel thread