LUA - ACU Hide Behavior flaw
The ACU Hide Behavior (Inspection Tour) was flawed - and would always send the ACU to the top left area of the base, especially very early in the game when there are no structures available for the code to 'run to'.  
Corrected this, and improved the routine to look for PD as well.   ACU will no longer wander far from the MAIN base centre core during the early game.