LUA - ParseIntel timing change - ratio calcs
Had recently moved to a much more aggressive timing cycle of 60 ticks - down from 75 - but it may be a bit too much - pushed it back up to 66 to relax it a bit.

PARSEINTEL now directly uses the new ratio process so retiring the old ratio calculation based on IMAP values - for AIR and LAND - Naval will come later.