LUA - Solving 'threat decay'
Threat decay is something that happens when a mobile unit 'dies'.   Unlike structures, when a mobile unit dies, it's threat values don't disappear from IMAP immediately, but persist for 3 minutes.   They stay at full strength for 30 seconds, and then decline over a 2 1/2 minute period.    
This is directly opposed to the way that threat moves instantly from one IMAP block to another.   If threat can move instantly - why isn't it removed instantly as it is for structures ?
The goal of this work is to eliminate that, so that all threat types function the same way.
In order to make this happen - we intercept the OnKilled event for all mobile units - and we add an entry into IMAP which reduces that threat to zero.  However, the original event persists - and then starts forcing IMAP into negative values.  To compensate, we re-add minute amounts of threat back to the IMAP table over the next 2 1/2 minutes to keep the net effect to near zero.
This necessitated some other related to changes, since we discovered that air units were, in the event that they 'crashed' - sidestepping the MobileUnit.OnKilled event (which was a bad thing).
As a result of this investigation, LOUD will no longer suppress low IMAP values, since this was contributing to the negative IMAP values, which, in turn, were causing some pathfinding issues.