LUA - pathfinding tweaks
One problem that pathfinding has always had concerns the way that the starting and ending nodes are located - it was always ONLY distance based - and did not take into account blocking terrain.  
We now check starting and ending nodes for blocking terrain elevation changes - and I've removed the shortcut in the pathfinding that would shortcut the final step based only on distance.
This should alleviate a great many clumping issues with cliffs and rivers.