LUA - AI Intel Ratios
Changed the basis of the calculation from 'me' vs. 'Avg Enemy TeaM' TO 'myteam' versus 'enemy'.   We'll continue to monitor this - still considering to make it 'myteam' versus 'currentenemyteam', but not sure I can identify specific teams yet - and more to the point ? Should it matter ?   I think it may - specifically if we want LOUD to focus on killing his current enemy.  Good questions - all.