LUA - Weapons DPS work
While going over DPS calculations with Phoenix, discovered that the FIRING CLOCK was in the wrong place in the sequence of events - coming after the actual firing had completed.  
The correct place is at the START of the firing sequence, so it now accurately shows how much time is left before the next firing event (provided there are not Rack reload delays and timeouts - which is does not account for - and never has).
The thread also never reached 100% completion, which has also been fixed.