LUA - AI Engineer tasks
The two most recent engineer factory assist tasks were not properly recognizing that there were no factories of the type is was trying to assist.  This was resulting in engineers at a naval base - trying to assist LAND or AIR factories at that base - which of course, would never happen.