LUA - AI BuildConditionMonitor changes
The BCM has always lengthened it's loop time based on the total number of conditions.   This results in a VERY long execution period as an AI adds more bases.  This, in turn, impacts how often a base can form platoons - and sometimes this delay is nasty - more than 5 MINUTES before a base recognizes it can form a platoon.
NOW - we bases the BCM loop length on the number of conditions we actually checked - and as a result - we can keep the BCM running every minute - and thus all the bases can form new platoons more frequently.
There is, of course, a performance cost associated with this - BUT - I can see now there are quite a few conditions that could be STATIC rather than instant conditions - moving the load off of the PFM (which is where we take a real performance hit from all the checks) - and onto the BCM.   In summary - these conditions get checked once every 60 seconds, instead of multiple times, at multiple bases.