LUA - AI BuilderManager cycle
Based on what we did to the BuildConditionMonitor cycle time, we can now change the relationship of the  BuilderManager in terms of how quickly it cycles - which really only applies to the PFM (since Factories and Engineers are self-directed).
We now run the primary bases at parity with the BCM, and other bases at half that speed.
This makes them MUCH more responsive since the BCM delay times are so much less than they had been.
This opens the opportunity to use more STATIC (non-instant) conditions as 'leading indicators' to prevent the PFM from executing condition checks multiple times across multiple locations.