LUA - AI - use new STATIC build conditions
With the recently added new STATIC build conditions, we convert the old 'per location' checks to these.
We also add new 'leading indicators' to the NAVAL PLATOON formers.   This essentially shuts down the checking of certain formations unless the entire POOL has the condition.    For example - for those formations that require 6 submarines, we insert the STATIC condition - and if false, none of the naval bases will make all the other UCBC checks usually associated with forming that platoon.