LUA - BCM optimization
The BuildConditionMonitor is the central point the controls the rate of platoon formation.  As a result, its cycle time can have a dramatic impact on overall game performance.   The changes here try to take into account the growing number of conditions that must be checked - gradually increasing the amount of time required to process them - but it must also be  aware of the fact that the PFM at every base has it's own conditions that must be checked.   Therefore it needs to be aware of - and adjust for - the number of production bases that each AI has.  It must also have some adjustment based upon the number of players (both human and AI) that might be in the game.

Tuning this can have dramatic impact on the effectiveness of the AI - if they cycle is too long - then the AI will be sluggish in forming platoons - if the cycle is too short - then overall performance will drop.