LUA - BuilderManager optimization
As described, the BCM generates the basic timing length of condition checking.   However, the 3 manager types, EM (engineers), FBM (factories) and the PFM (Platoon Formation) all use this data to control their own processes.   Engineers and Factories are self-directed, and aren't really impacted by this cycle - as they check conditions only as needed.    The PFM is, however, directly affected by the cycle time of the BCM - so we need to optimize it, in relation to the changes we made in the BCM.

Every base has a PFM - and while it would be nice if we could run them all as fast as the BCM, that really isn't practical - too many conditions and performance would drop rapidly as the AI expands.   So - we use the concept of PRIMARY bases - since this is where most of the units will be - we run these PFMs at about half the speed of the BCM - all other bases run at about 1/5 of that speed.

These values represent a reasonable compromise between responsiveness and performance.
Optimized this by making sure we only resort the builders - and process all the conditions - IF THERE ARE UNITS at that base.

Optimized the checking of LAND versus AMPHIB modes so that a check is only done if the AttackPlan Goal has changed.

GetHighestBuilder - used frequently by engineers and factories -