LUA - AI platoon move - slackdistance
Slackdistance controls the distance from a waypoint that a platoon needs to be before executing the next leg in a path - naval platoons were getting the original default of 16, which was often not enough for large naval formations to realize they were at the waypoint.

Gave Naval movement specific slackdistance of 28 - also, moved the general default from 16 up to 20.