ENHANCEMENT MODS - Added to Git
Updated and revised - tamed the Enhanced Commanders mod to more reasonable levels to address the problems it can cause the the AI's opening economy - and the OP nature of the mod in general.

Updated the Enhanced Artillery to simplify it's structure - removed duplicated data, same with the Commanders mod.

Updated version to 1.2