UNITS - Reversed Civilian Units split
Due to a great number of unspecific errors I have reverted this project for the time being.   A larger study needs to be made of the interactions involved.

While I appreciate the some 20MB memory difference this can make, at this time, I don't have the time to resolve all the issues raised by it.