LUA - AI Mult now impacts MassPointShare
MassPointShare, and it's variants, are used to tell the AI when he has his 'share' of mass control.

By applying the AI Mult to it, we can flip him over in to an aggressive stance earlier than normal.

We are NOT, however,  taking into account the relative balance in team sizes yet.

Also, adjusted the VeterancyMult value to insure it's never below 1, which only impacts AIMult when it's below 1.