LUA - AI fix for T4 Air units idle
When damaged below 50%, all air units go into the airrefit thread which is intended to return the unit to an airpad for repair - and refuel.
Previously, T4 air units were exempt from this, but we now have them RTB for repair, to increase their survivability.   However - once in the refit thread, the T4 Air units would never exit - as they don't have fuel.

The code was specifically looking for both health AND fuel ratio to release a unit from refit.   It now is aware that a unit may not have a fuel component.