LUA - Threat/Attack Planner changes
The Attack Planner's intention is to develop a 'goal' - and that goal is used to determine the placement of primary bases.   However, the mechanism was always defaulting to the enemy start position - and not the most valuable, distance evaluated target.
This update changes this - making closer dangerous targets become the goal, and thus changing the placement of 'PRIMARY' bases.
This is intended to deal with 'bypassing' situation - and keep LOUD focused on clearing dangerous positions closer to home first.