BrewLAN - vers 0.8 for LOUD
This update to BL is filled with all kinds of minor updates, mostly to textures, icon and minor visual things.

There are however, a few new additions, such as the UEF Penetration Fighter, and the SERA Sniper Experimental.

There are also some unit model revisions, notably the AEON mobile radar and the CYBRAN mobile AA

Included is also  a number of stat tweaks across a broad number of units.