LUA - Attack Planner work
Additional development for the AI attack planner.  In it's original form, it simply plotted a path from MAIN to the home of it's selected enemy.
It's now updated to make plans to the most threatening targets, revised every 8 minutes, and point choices will now take into account pathing lengths as part of the selection process.
Also, the plan will now reliably indicate if it's an Amphibious or Land plan, and trigger the correct production choices accordingly.
The final point is no longer the marker closest to the goal, but the actual threat valued goal position.