AI - trigger point to exit from 'Skirmish'
Skirmish mode is the condition the AI is in at the start of the game - based upon mass point ownership.   Previously - the AI would seek to own his 'share' of the mass points before moving into an attack posture.  
This is now modified by the AI Multiplier, so a cheating AI will flip over to attack posture sooner depending upon the multiplier.