LUA - Platoon Formation Manager cycle time
The PFM is responsible for forming platoons out of units at a base.   It runs on a cycle that is based upon the length of time the AI requires to process ALL the static build conditions.  
Previously the PFM ran at HALF that speed, resulting in conditions where it might be 2.5 minutes between formation of new platoons.    This was too long.   
The PRIMARY base will now run at full speed, with an adjustment for the AIMultiplier.  The results in dramatically less time between formation creations - allowing the PRIMARY to make far more immediate use of the resources at hand.