LUA - AI - Engineer Suppression
One of the hallmarks of aggressive human play is the early explosion of undefended engineers out to exploit resources as quickly as possible.   Previously, LOUD was NOT really focused on those engineers.   This changes the focus, primarily of early air groups, but introduces a new early land formation specifically to deal with this.