LUA - AI - Base Distress Response
Two minor changes to base responses - first, the response threshold is moved up to 65% of the enemy threat.
The second change is to the amount of delay between movement orders.   An increased delay is intended to tighten the spacing between responders by allowing those further away, more time to get moving before those units which may be closer.