LUA - AI tweaks to Skirmish mode land formations
Creation and tweaks of new Extractor attack formations, both for the early phase of the game - one is similar to early Mass Point and MEX guads, but offensive in nature.    The other, slightly larger, forms a little bit later when production is higher, to maintain pressure of enemy resource collection.