LUA - SHIELD
Made the shield status obligatory, as shields that were in recharge state could come back on after being loaded onto transport.  This prevents that behavior by setting the Shield to an ON condition when in recharge mode, which is technically correct.   This allows the transport loading code to properly disable the shield.