LUA - Map Wrecks - decay now 30 minutes
Normally, wrecks will despawn after 15 minutes, but wrecks that are created as map objects, will get a 30 minute timer.   This is done by adding an optional overridetimer value when invoking the CreateWreckageProp function.

Various EFFECTUTILITIES now track and return the emitters they create, for use in trash handling (if that's being done during the prop creation).