LUA - AI Factory counts decreased
In ALL production bases, factory counts have been decreased.   The effect of this change is an earlier transition into T4, with a more supportable economy.  A side effect of this change will possibly be the more urgent desire to expand, since there may be a greater degree of surplus resources, and/or base defenses will be more aggressively pursued.