LUA - AI - FindTargetInRange (NavalForceAI)
This is a basic target finding task - given target categories and a range for a specific squad of units to find a target.   It now can check for blocking terrain if the LayerCheck flag is true.
NavalForceAILOUD didn't use to check this - now it does, so it won't try to path over obstructions or target land positions.
Also - adjust searchradius for all naval platoons down to about 180-220 range.