LUA - AI EM Base response radius
Noticed that base responses by one base could draw units in from another nearby base - so naval units would get pulled in from a nearby Naval position, and be told to RTB to the land base.
Base Repsonse draw radius is no longer defaulted to 200, but is tied to the normal formation radius of that base.   This should eliminate interactions like that.