LUA - AI - ACU starts LIFETHREAD
This is a very subtle AI cheat that keeps the AI from crashing his eco by granting him tiny amounts of Mass & Energy - but only when his storage is empty - and his  demand exceeds supply.    Will return him to a zero status - so this is driven entirely by construction shortfalls, but does not give him the resources required to start new projects - so it's very very subtle.

At the moment, it's tied to the ACU - so if he dies - the bonus goes with it.