Units - T3 Aeon Transport
The Teleportation Feature of the T3 Aeon Transport is one that is from Blackops. This was a problem as it was the instant teleportation which basically required anti teleport which is not in Default LOUD, this transport even against Anti Teleport is extremely cheesy. To fix this Teleportation has been removed from transport but instead replaced with a Minor Speed Buff and general Ability Buff.