Some corrections to Seraphim protector bot .bps
The Seraphim have a lineup of protector bots running from T1 to T4, all of which had incorrect icons, except the T4, which just had a description typo. This same typo was fixed in most of the other blueprints.

It may be vaguely beneficial to the UX to give these bots discrete icons from direct-fires and shields, but nobody has mentioned it yet so it's probably not a big deal.