LUA - AI Naval Base creation
Naval Bases were suppressed if a major enemy threat was within 13km.   On some maps, this is pretty much guaranteed - such as Lily Pads.   This feature was originally intended to keep an AI out of the water if the enemy was close by.  

In that respect - it is valid (it was created to help with Seton's Clutch situations) - but it's to broad.  Reduced the range to 12km (600) and increased the threat trigger to 1500.   This should permit artificial threat (at game start) to decline enough, at a soon enough point, to allow naval bases to form.