Seraphim protector unit polish
I noticed that the Sera T2 Restoration Field Generator looks exactly identical to a Sera T2 radar from a strategic zoom level, and figured that if I was going to give it a unique icon to improve readability, I would do the same for the rest of the Sera "Protector" class units.

The kite shield motif seemed suitable; the Red Cross, Star of Life, and Staff of Asclepius all fail to translate nicely to a strategic icon.

I also cleaned their ability lists and descriptions, in addition to fixing the icon on the similar Shielded Protector Bot brought in by WBP.