Reduce ratio ACT start delay to 5 minutes
I Initially used a timer of 7 minutes to prevent the ratio adaptive from starting to work before the AI started detecting enemy units and creating a sane land ratio. Thinking more about this, I realize that a skilled player could probably aggress on an AI in multiple fewer minutes than that on a 10x10 map. 5 minutes should be a happier medium.