Force ACT lobby options to bottom of options lists
Axle reported getting overwhelmed when first encountering all of the lobby settings relating to ACT. He was likely not the first and (this commit notwithstanding) he will likely not be the last.

These settings for intermediate-level users of LOUD will now sink to the bottom of the lobby settings list, and to the bottom of the map selection settings panel. Users interested in scanning through all of the options will find them and be free to explore them at their own leisure.