Fix ratio ACT to be consistent with timed ACT, improve logging
Because ratio ACT worked off of the AI's baseline cheat multi previously, it would not have a consistent effect on the army pool buff if the timed ACT was simultaneously active.

Additionally, logging of the ratio ACT kickoff was wrong in taking place before the initial delay finished. There is now a message indicating that the ACT will start after 5 minutes and another when that actually takes place, which also lists the parameters of the adaptation.

Other polish to logging for both forms of ACT has been included.