LUA - AI Base Expansions
The parameters around expansions (Production and DP) have been tweaked to make LOUD a bit more aggressive with regards to getting out on the map.   A lot of the parameters across the types have been harmonized, and a slightly greater priority is given to production expansion when the conditions are met.