LUA - Base Distress Monitor
I've revamped some parameters around the Base Distress Monitor, to improve it's reaction time - which in turn allows it to relieve itself of the alert a little more quickly when the alert is gone.   If a base cannot respond to an alert, the alert will cancel itself after several tries.   
Also, Base Alerts will prevent the Dead Base Monitor from closing a base until the alerts have expired.   This was causing a fair number of alert units to get 'stuck' at a base that was closed during an alert.