LOUD Standard modlist button
A button in the upper section of the Mod Manager GUI now enables all of the mods we use as standard in our community games: the unit packs, EvenFlow, Integrated, Unit Additions, Structure Enhancements, BOACUs, and all of the UI mods. Barista is included here because I believe it ought to be part of the standard and is harmless.

A new button texture was included so that the Save and Load buttons could be shrunk, allowing the new button to fit in the appropriate place.

I also linted modmanager.lua while I was at it.