Map select list folder grouping
The map selection list will now collect maps into expandable/collapsible folders.

Random selection does not work yet, and folders are not hidden if all of their contents are filtered out. There is also a substantial delay every time a folder is interacted with, since the list needs to be entirely repopulated.