Lobby field to set all AI multipliers
A textbox next to the "Add AIs" and "Clear AIs" buttons now acts as a global setting for new AI multipliers. Any AI added either via the Fill button or via an individual slot will take on this multiplier setting. There is also a "Set All AI Cheat" button next to the new textbox that will propagate that setting to any existing AI players in the lobby.