Unit Cap lobby setting text box
This commit lays the infrastructure for any lobby option to be set via an edit box. It has not been stress-tested, and there is no way to validate input for sanity yet. There is also no control on the sim side to check if the unit cap is high enough for a reasonable game to play out.