LUA - AI - CustomUnits (Land Units)
Custom Unit Building Review Summary:

Some Major Fixes here are in the Amphibious Section Where LOUD can now build T2.5 Amph Units. 
----
General Review and lowering of Units with Certain Footprint Issues and Units that would generally be better for the AI to use and build.
-----
Updated some # to -- for Commented Out Sections
-----
Heavier T3.5 was generally buffed in Build Chance 
Lighter T3s were Lowered and Rised in certain sections to address Mixed Bag Platoons. Some Lighter T3s were traded out for the AI due to effectiveness, abilities, or function.
-----
T1.5 was rised in-general as we make a lot of earlier T1 that tends to be more of the lower tier of T1.
-----
T4s were priorities were switched around a little in certain Areas
------
Note: No Unit was removed for the AI, Only Lowered or Rised in desire to build such.