LUA - HOTBUILD keybinding et al.
This is a working version of HOTBUILD - but with some features incomplete.   Buildmode is the default - with B - and the Keylabels for each item will appear on the Avatars.

To ENABLE HOTBUILD - you must bring up the keymapper (F1), and RESET your keybindings, an option will appear during this sequence to prompt you to configure ZULANS HOTBUILD - and a language preference to go with it.
At this point, any bound key used by HOTBUILD will initialize it (for example W is used to build a factory).
There will be a measurable pause while all the keybinds are read - but just that one time.
Now - as of yet - I have not located the data or the means by which we can replace the original ID letters that appear in the builds - and if I turn them on, the original BUILDMODE key assignments will show - which is wrong.   For now - I simply don't turn them on.

In summary, HOTBUILD is working, but there are currently no visual cues.

BE ADVISED - HOTBUILD remaps core function keys like F for Ferry - and this is the importance of the visual cues.   

How does FAF do it.  Technically they don't.  They've completely removed BUILDMODE by unmapping the B key (trigger for BUILDMODE).    HOTBUILD is enforced on all users.