LUA - AI Brain - framework for Prebuilt units
I have some changes coming to pre-built units (Menu Option), that will permit the spawning of particular units, as part of the pre-built units code.  For now, those just happen to be BO ACU's - but the code does not detect the mod being installed and just bypasses it for now.

The true change, that does work, is the logic that merges that data, with the original data - thus enabling stacking of the data sets.