Basics of Unit DB dialog
This is the first commit in yet another skunkworks project, this time to create an in-game unit database. The plan is to make it accessible from both the main menu and the sim, and allow searching through and viewing entries based on all of the BPs from all of the LOUD-standard mods. Whether or not any of this is even viable is TBD.