UNIT DB: Trim Civilian, Insignificant, and Benign units
These tend not to have icons and aren't very useful to the player (at least not in isolation).

This adds a large branching statement to every unit blueprint parse, but the first-time build of the DB still runs pretty fast in the production build on my AMD FX 8350, so I'm not fretting over this for now.