LUA: Fix invalid keybinding sort function
Sorting of the keybinding menu relies upon key action description strings to create an alphabetically-ordered list; if any key does not have a keyDescription entry, the order of the table.sort() arguments becomes moot and the function fails.

I don't expect this to be a fix for all of the errors which have been reported recently related to opening the keybinding menu, but this probably helps somebody.

On the way to this fix I tidied up keyactions.lua and corrected a typo.