BrewLAN: Add Lucky Dip
Balthazar suggested his mod Lucky Dip as something of a solution to unit redundancy. The more I thought of this idea, the more I liked it, and there are no problems which might hinder its integration.

At the moment, it is more-or-less in its original state. I accounted for the differences in unit roster between FAF BlOps and LOUD BlOps, and improved the commenting.