LUA - SIM - EM Base Response
This is a general review in my eyes of Base Response, Watching about 6 replays of mostly players rushing the AI, and watching AI Fight eachother. I came to this changes stated in the EM.Lua. 

First was expanding the radius from 90 to 120 will allow him more accurate Threat Judgement. During debugging I noticed he was definitely missing out on simply a larger radius to search as players would not often condense in such a small radius.

The biggest one is the Alerts Table, I noted in the EM.Lua that Sprouto's Comments are noted in terms of Ping Ponging between targets. I did some extensive testing on some very rough replays for the AI in terms of him being surrounded in one base, didn't notice much of this ping-ponging but ill be watching in more games. Anyway, The Alert Tables can now log multiple of the Same Threat Types which means he can now recognize if there are 2 different armies on both sides. 

Another Change was upping both Naval and Land Base Responses to muster at least 70% of the Enemy Threat before Responding. 

The Last Change is Distress Repeats, now this actually was recently changed in the last few months by Sprouto. Now i am reviewing this, I up 6 ticks to 8 ticks too allow a much stronger wave to come out of the base, From 6 to 8 was actually rather noticable in a lot of the replays.