UNIT DB: Hotkeys for scrolling, displaying units
The up and down arrows can now be used to scrub through the unit list. Adding Shift will scroll by 10 at a time, and Ctrl will scroll to the top or bottom (the last feature is unconfirmed to be working properly, since an unrelated bug seems to be happening when hitting the bottom of the list).

In support of this feature, unit list entries now store their unit index, and the DisplayUnit() and FillLine() functions don't take BPs or IDs as args anymore, since they can be retrieved easily given the unit index.