LUA - Naval base templates
As we introduce 10k water maps, we're faced with some hard choices about Naval bases, since 10k maps rarely have an area large enough for all but the TINY (2 factory) naval layout.

There had been a logic block that was preventing the selection of ANY naval base on a 10k map - but this change should rectify that.

The ECO threat range restriction was lowered from 12k to 9k to permit Naval bases to happen on such small maps (which are usually directly connected).